requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0
channels==3.0.3
channels-redis==3.2.0
psycopg2-binary==2.8.6
coverage==5.5
